Simple gRPC interface on teh Tbl service

+++ src/orders/Orders/PlaceOrder/PlaceOrderRequest.cs
@@ -10,13 +10,13 @@ namespace Orders.Orders.PlaceOrder
         int? Id { get; set; }
 
         /// <summary>Order size.</summary>
-        decimal? Amount { get; set; }
+        double? Amount { get; set; }
 
         /// <summary>The direction of the order; buy or sell.</summary>
         BuySell? BuySell { get; set; }
 
         /// <summary>Order Price. Optional for market orders.</summary>
-        decimal? OrderPrice { get; set; }
+        double? OrderPrice { get; set; }
 
         /// <summary>Order type.</summary>
         PlaceableOrderType OrderType { get; set; }
@@ -28,13 +28,13 @@ namespace Orders.Orders.PlaceOrder
         OrderDuration OrderDuration { get; set; }
 
         /// <summary>Stop limit price for Stop Limit order.</summary>
-        decimal? StopLimitPrice { get; set; }
+        double? StopLimitPrice { get; set; }
 
         /// <summary>Distance to market for a trailing stop order.</summary>
-        decimal? TrailingStopDistanceToMarket { get; set; }
+        double? TrailingStopDistanceToMarket { get; set; }
 
         /// <summary>Step size for trailing stop order.</summary>
-        decimal? TrailingStopStep { get; set; }
+        double? TrailingStopStep { get; set; }
     }
 
     /// <summary>Request contract for "Place order"</summary>
@@ -45,11 +45,11 @@ namespace Orders.Orders.PlaceOrder
         /// <summary>Unique id of the instrument to place the order for.</summary>
         public int? Id { get; set; }
         /// <summary>Order size.</summary>
-        public decimal? Amount { get; set; }
+        public double? Amount { get; set; }
         /// <summary>The direction of the order; buy or sell.</summary>
         public BuySell? BuySell { get; set; }
         /// <summary>Order Price. Optional for market orders.</summary>
-        public decimal? OrderPrice { get; set; }
+        public double? OrderPrice { get; set; }
         /// <summary>Order type.</summary>
         public PlaceableOrderType OrderType { get; set; }
         /// <summary>The Instruments AssetType.</summary>
@@ -59,11 +59,11 @@ namespace Orders.Orders.PlaceOrder
         /// <summary>Used when placing related orders for an existing position.</summary>
         public string PositionId { get; set; }
         /// <summary>Stop limit price for Stop Limit order.</summary>
-        public decimal? StopLimitPrice { get; set; }
+        public double? StopLimitPrice { get; set; }
         /// <summary>Distance to market for a trailing stop order.</summary>
-        public decimal? TrailingStopDistanceToMarket { get; set; }
+        public double? TrailingStopDistanceToMarket { get; set; }
         /// <summary>Step size for trailing stop order.</summary>
-        public decimal? TrailingStopStep { get; set; }
+        public double? TrailingStopStep { get; set; }
 
         public PlaceRelatedOrder[] Orders { get; set; }
         public OpenOrderRelation? OpenOrderRelation { get; set; }
@@ -93,11 +93,11 @@ namespace Orders.Orders.PlaceOrder
             /// <summary>Unique id of the instrument to place the order for.</summary>
             public int? Id { get; set; }
             /// <summary>Order size.</summary>
-            public decimal? Amount { get; set; }
+            public double? Amount { get; set; }
             /// <summary>The direction of the order; buy or sell.</summary>
             public BuySell? BuySell { get; set; }
             /// <summary>Order Price. Optional for market orders.</summary>
-            public decimal? OrderPrice { get; set; }
+            public double? OrderPrice { get; set; }
             /// <summary>Order type.</summary>
             public PlaceableOrderType OrderType { get; set; }
             /// <summary>The Instruments AssetType.</summary>
@@ -105,10 +105,10 @@ namespace Orders.Orders.PlaceOrder
             /// <summary>The Order Duration.</summary>
             public OrderDuration OrderDuration { get; set; }
             /// <summary>Stop limit price for Stop Limit order</summary>
-            public decimal? StopLimitPrice { get; set; }
+            public double? StopLimitPrice { get; set; }
             /// <summary>Distance to market for a trailing stop order.</summary>
-            public decimal? TrailingStopDistanceToMarket { get; set; }
+            public double? TrailingStopDistanceToMarket { get; set; }
             /// <summary>Step size for trailing stop order.</summary>
-            public decimal? TrailingStopStep { get; set; }
+            public double? TrailingStopStep { get; set; }
         }
 }
